+++ lib/services/g-office-list-domain.dart
@@ -4,9 +4,9 @@ import 'package:laokyc_button/constant/api_path.dart';
 import 'dart:convert';
 import 'package:laokyc_button/model/list_domain_model.dart';
 import 'package:laokyc_button/widgets/dialog_loading.dart';
+import 'package:laokyc_button/widgets/error_dialog.dart';
 
-Future<ListDomainModel> listDomain(BuildContext context) async {
-  ListDomainModel dataFromAPI = ListDomainModel();
+Future<ListDomainModel?> listDomain(BuildContext context) async {
   String url = APIPath.listDomain;
   showDialog(
       context: context,
@@ -15,15 +15,66 @@ Future<ListDomainModel> listDomain(BuildContext context) async {
       });
   try {
     var response = await http.get(Uri.parse(url));
-    Navigator.pop(context);
     if (response.statusCode == 200) {
-      dataFromAPI = ListDomainModel.fromJson(json.decode(response.body));
+      Navigator.pop(context);
+      return ListDomainModel.fromJson(json.decode(response.body));
     } else {
-      throw ('fail');
+      Navigator.pop(context);
+      errorDialog(
+          context,
+          'ຂໍອະໄພ',
+          'List domain: ${response.statusCode} ${response.body}',
+          'ປິດ',
+          'Phetsarath');
+      return null;
+    }
+  } catch (e) {
+    return await listDomainExceptionOne(context);
+  }
+}
+
+Future<ListDomainModel?> listDomainExceptionOne(BuildContext context) async {
+  String url = APIPath.listDomain;
+  try {
+    var response = await http.get(Uri.parse(url));
+    if (response.statusCode == 200) {
+      Navigator.pop(context);
+      return ListDomainModel.fromJson(json.decode(response.body));
+    } else {
+      Navigator.pop(context);
+      errorDialog(
+          context,
+          'ຂໍອະໄພ',
+          'List domain: ${response.statusCode} ${response.body}',
+          'ປິດ',
+          'Phetsarath');
+      return null;
+    }
+  } catch (e) {
+    return await listDomainExceptionTow(context);
+  }
+}
+
+Future<ListDomainModel?> listDomainExceptionTow(BuildContext context) async {
+  String url = APIPath.listDomain;
+  try {
+    var response = await http.get(Uri.parse(url));
+
+    if (response.statusCode == 200) {
+      Navigator.pop(context);
+      return ListDomainModel.fromJson(json.decode(response.body));
+    } else {
+      Navigator.pop(context);
+      errorDialog(
+          context,
+          'ຂໍອະໄພ',
+          'List domain: ${response.statusCode} ${response.body}',
+          'ປິດ',
+          'Phetsarath');
+      return null;
     }
   } catch (e) {
     Navigator.pop(context);
     throw (e);
   }
-  return dataFromAPI;
 }