+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## 0.4.9
-* TODO: A
+* TODO: dialog loading